Estrutura de decisão https://wiki.python.org.br/EstruturaDeDecisao
Refatoração do exercício 01

+++ estrutura_decisao/exercicio_01.py
@@ -1,8 +1,8 @@
 """ 
 1. Faça um Programa que peça dois números e imprima o maior deles.
 """
-numero1 = int(input("Digite o primeiro número: "))
-numero2 = int(input("Digite o segundo número: "))
+numero1 = float(input("Digite o primeiro número: "))
+numero2 = float(input("Digite o segundo número: "))
 if numero1 > numero2:
     print(f"O maior número é: {numero1}")
 elif numero2 > numero1:
